Create Java Admin client module
