+++ langstream-runtime/langstream-runtime-impl/src/main/java/ai/langstream/runtime/agent/AgentCodeDownloader.java
@@ -58,6 +58,9 @@ public class AgentCodeDownloader {
     }
 
     private static void downloadCustomCode(CodeStorageConfig codeStorageConfig, Path codeDownloadPath) throws Exception {
+
+
+
         if (codeStorageConfig != null) {
             log.info("Downloading custom code from {}", codeStorageConfig);
             log.info("Custom code is stored in {}", codeDownloadPath);
